- Update window ui widget, add close icon, fix bugs and dimensions. - Fix missing return reference on getText() method. - Add new Table widget for displaying nice multidimensionnal datas. - Add new Menu widget with icons and shortcut hints. Only one level of depth for now. - Implement the basic list component with same principles of menu & table. - Improve layout widget and add debug draw method. - Improve and fix bugs in the input widget. Also add new icon for resetting the field. - Add new default icons into the icon widget. - Implement new checkbox widget. (Starting) - Minors update on the button widget. - Update the Rect shape class to add utility methods. - Minors update on the WidgetState class - Add connections to the main widget class with the table & Menu widgets. - Prepare the theme class. - Start impementing the ui layer manager ! - Refactor the guiManager and add minors update (facilitate dependencies, new getWidgetByName method etc...); - Try some examples with the quadtree implementation in the gui (to remove later !!) - Update the renderer class to be able to scissors the viewport and properly integrate it with the ui. - Add new demo for the recent stuff added in the main game class. - Implement the starting of a quadtree ! - Add new mousewheel event support to the inputManager class ! - Handle the mousewheel event on the FPSCamera class to control its speed ! - Fix bug on the assetManager which wasn't returning the correct value. - Add new nlohmann / JSON lib to the include files. - Add new demo meshes ! - Also a lot of little bug fixing, code formatting & arrangements !

+++ PXL/src/assets/loader.cpp
@@ -12,7 +12,7 @@ Loader::Loader()
 Mesh* Loader::loadToVAO(const std::string& name, std::vector<glm::vec3> vertices)
 {
 	GLuint vao = this->createVAO();
-	size_t size = vertices.size();
+	unsigned int size = vertices.size();
 
 	this->storeDataInAttributeList(0, 3, &vertices[0], size * sizeof(glm::vec3));
 	this->unbindVAO();
@@ -23,7 +23,7 @@ Mesh* Loader::loadToVAO(const std::string& name, std::vector<glm::vec3> vertices
 Mesh* Loader::loadToVAO(const std::string& name, std::vector<glm::vec2> vertices)
 {
 	GLuint vao = this->createVAO();
-	size_t size = vertices.size();
+	unsigned int size = vertices.size();
 
 	this->storeDataInAttributeList(0, 2, &vertices[0], size * sizeof(glm::vec2));
 	this->unbindVAO();
